pyproject.toml
[project]
name = "pythonrestapi"
version = "1.0.0"
description = "Python RestAPI using Quart HTTP/3 ASGI framework."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marshmallow",
    "python-dotenv",
    "Quart-CORS",
    "PyJWT",
    "alembic",
    "itsdangerous",
    "quart",
    "quart-wtforms",
    "aioquic",
    "quart-trio",
    "jsonpickle",
    "quart-bcrypt",
    "quart-keycloak",
    "hypercorn[h3,trio]",
    "uuid7",
    "google-genai",
    "quart-uploads",
    "flask-sqlalchemy",
    "quart-flask-patch",
    "psycopg[binary,pool]",
    "marshmallow-sqlalchemy>=1.4.2",
]

[dependency-groups]
dev = [
    "pytest-flask",
    "pytest",
    "pytest-asyncio",
    "beautifulsoup4",
    "tox",
]
